+++ backend/server/restservice/subnets_test.go
@@ -1604,7 +1604,7 @@ func TestCreateSubnet4BeginSubmit(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -2097,7 +2097,7 @@ func TestCreateSubnetBeginCancel(t *testing.T) {
 	rsp2 := rapi.CreateSubnetDelete(ctx, params2)
 	require.IsType(t, &dhcp.CreateSubnetDeleteOK{}, rsp2)
 
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -2553,7 +2553,7 @@ func TestUpdateSubnet4BeginSubmit(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -2871,7 +2871,7 @@ func TestUpdateSubnet4BeginSubmitChangeSharedNetwork(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -3067,7 +3067,7 @@ func TestUpdateSubnet4BeginSubmitAddToSharedNetwork(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -3261,7 +3261,7 @@ func TestUpdateSubnet4BeginSubmitRemoveFromSharedNetwork(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -3592,7 +3592,7 @@ func TestUpdateSubnet6BeginSubmit(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -3815,7 +3815,7 @@ func TestUpdateSubnet6BeginSubmitRemoveFromSharedNetwork(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -4012,7 +4012,7 @@ func TestUpdateSubnet6BeginSubmitAddToSharedNetwork(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -4491,7 +4491,7 @@ func TestUpdateSubnetBeginCancel(t *testing.T) {
 	rsp2 := rapi.UpdateSubnetDelete(ctx, params2)
 	require.IsType(t, &dhcp.UpdateSubnetDeleteOK{}, rsp2)
 
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
